requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
jinja2==3.1.2
requests==2.31.0
httpx==0.25.2
black==23.12.0
autopep8==2.0.4
isort==5.13.2
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
# For async HTTP
aiohttp==3.9.1
# Optional: for better logging
python-json-logger==2.0.7
# Optional: for database (Phase 3)
sqlalchemy==2.0.23
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

